refect: enum 클래스 도메인으로 추출
